[fedora-toolbox] push latest toolbox images to registries

+++ scripts/sync-latest-container-base-image.sh
@@ -65,6 +65,8 @@ fi
 #
 build_name=$(koji -q latest-build --type=image f${1}-updates-candidate Fedora-Container-Base | awk '{print $1}')
 minimal_build_name=$(koji -q latest-build --type=image f${1}-updates-candidate Fedora-Container-Minimal-Base | awk '{print $1}')
+toolbox_build_name=$(koji -q latest-build --type=image f${1}-updates-candidate Fedora-Container-Toolbox | awk '{print $1}')
+
 
 if [[ ${1} -eq "$current_stable" ]]; then
     tagname="latest"
@@ -118,6 +120,7 @@ if [[ ${1} -gt "$current_rawhide" ]]; then
     exit 1
 fi
 
+# For fedora-container-base
 if [[ -n ${build_name} ]]; then
     # Download the image
     work_dir=$(mktemp -d)
@@ -135,6 +138,7 @@ if [[ -n ${build_name} ]]; then
     printf "Removing temporary directory\n"
     rm -rf $work_dir
 fi
+# For fedora-container-minimal-base
 if [[ -n ${minimal_build_name} ]]; then
     # Download the image
     work_dir=$(mktemp -d)
@@ -148,7 +152,23 @@ if [[ -n ${minimal_build_name} ]]; then
     popd &> /dev/null
 
     generate_manifest_list fedora-minimal ${1}
+    printf "Removing temporary directory\n"
+    rm -rf $work_dir
+fi
+# For fedora-container-toolbox
+if [[ -n ${toolbox_build_name} ]]; then
+    # Download the image
+    work_dir=$(mktemp -d)
+    pushd ${work_dir} &> /dev/null
+    koji download-build --type=image  ${toolbox_build_name}
+    # Import the image
+    for arch in "${ARCHES[@]}"; do
+        xz -d ${toolbox_build_name}.${arch}.tar.xz
+        copy_image docker-archive:${toolbox_build_name}.${arch}.tar fedora-toolbox:${1}-${arch}
+    done
+    popd &> /dev/null
 
+    generate_manifest_list fedora-toolbox ${1}
     printf "Removing temporary directory\n"
     rm -rf $work_dir
 fi
